Refatorar método atualizar status na Gestor Controller.
