+++ src/main/java/br/com/zup/PointMarker/funcionario/FuncionarioController.java
@@ -1,22 +1,8 @@
 package br.com.zup.PointMarker.funcionario;
 
-import br.com.zup.PointMarker.enums.Status;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarCargoDTO.AtualizarCargoEntradaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarCargoDTO.AtualizarCargoSaidaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarSalarioDTO.AtualizarSalarioEntradaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarSalarioDTO.AtualizarSalarioSaidaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarStatusDTO.AtualizarStatusEntradaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.AtualizarStatusDTO.AtualizarStatusSaidaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.CadastroFuncionárioDTO.CadastroFuncionarioEntradaDTO;
-import br.com.zup.PointMarker.funcionario.dtos.CadastroFuncionárioDTO.CadastroFuncionarioSaidaDTO;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/funcionario")
@@ -30,14 +16,5 @@ public class FuncionarioController {
     public FuncionarioController(FuncionarioService funcionarioService, ModelMapper modelMapper) {
         this.funcionarioService = funcionarioService;
         this.modelMapper = modelMapper;
-    }
-
-    @PutMapping("/status/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public AtualizarStatusSaidaDTO atualizarStatus(@PathVariable int id, @RequestBody AtualizarStatusEntradaDTO atualizarStatusEntradaDTO) {
-
-        Status status = modelMapper.map(atualizarStatusEntradaDTO, Status.class);
-        funcionarioService.atualizarStatus(id, status);
-        return modelMapper.map(status, AtualizarStatusSaidaDTO.class);
     }
 }
